requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==3.0.4
discord.py==2.1.0
feedparser==6.0.8
fuzzywuzzy==0.18.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.2
PyYAML==6.0
rfc3986==1.5.0
sgmllib3k==1.0.0
sniffio==1.2.0
typing-extensions==4.0.1
yarl==1.7.2
